+++ mud-world-contract/src/main/java/com/cailleach/mudengine/world/rest/dto/PlaceExit.java
@@ -20,7 +20,7 @@ public class PlaceExit {
 	
 	private boolean lockable;
 
-	private Integer targetPlaceCode;
+	private Long targetPlaceCode;
 	
 	
 	public static String getOpposedDirection(String direction) {
